Revert "Import Cronet v136.0.7091.2 from Chromium to AOSP."
This reverts commit 759378822ded2bca748230396540c08165777bac.

Reason for revert: Broke git_main-without-vendor

Bug: 414389515

+++ stable/net/http/broken_alternative_services.h
@@ -174,8 +174,7 @@ class NET_EXPORT_PRIVATE BrokenAlternativeServices {
 
   struct AlternativeServiceHash {
     size_t operator()(const AlternativeService& entry) const {
-      return static_cast<size_t>(entry.protocol) ^
-             std::hash<std::string>()(entry.host) ^ entry.port;
+      return entry.protocol ^ std::hash<std::string>()(entry.host) ^ entry.port;
     }
   };
